% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rule_set.R
\name{create.rule.set}
\alias{create.rule.set}
\title{Take vector of elements and calculate probabilities of elements and combinations occurring}
\usage{
create.rule.set(elements, maxlen)
}
\arguments{
\item{elements}{list with vectors for all elements observed together at each event}

\item{maxlen}{maximum size of combinations to be considered}
}
\value{
Function returns a dataframe with observed probabilities for each combination in the dataset
}
\description{
Function underlying the netfacs() function; takes the observed elements, applies the apriori() function of the arules package, and create dataframe with all observed probabilities
}
