% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat2list}
\alias{mat2list}
\title{Turn matrix into list}
\usage{
mat2list(inmat)
}
\arguments{
\item{inmat}{numeric or integer matrix, 0/1 matrix. Must have column names!}
}
\value{
a list where each item represents a row in \code{inmat} and is a
        character vector with the column names for which a row value is 1
}
\description{
Turn matrix into list
}
\details{
The function will fail if the input matrix does not have column
         names or if it is a not a \code{matrix}, but for example a
         \code{data.frame}.
}
\author{
Christof Neumann
}
