% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shuffle_checkerboard}
\alias{shuffle_checkerboard}
\title{Randomize entries in matrix while keeping row and column sums fixed}
\usage{
shuffle_checkerboard(inmat, swaps = 1L, checkerboards_only = TRUE)
}
\arguments{
\item{inmat}{numeric or integer matrix, 0/1 matrix}

\item{swaps}{numeric, the number of (attempted/trial) swaps to perform
(default is \code{1})}

\item{checkerboards_only}{logical, should the random selection of trial
matrices to be swapped be constrained to checkerboards (default is
\code{TRUE})}
}
\value{
a randomized matrix with the same dimensions as \code{inmat}, where
row and column sums are constant
}
\description{
This is a utility function and as such probably not of much interest for a
typical user.
}
\details{
The algorithm starts by selecting a random submatrix with 2 rows
         and 2 columns (trial matrix). An actual swap is performed if this
         selected submatrix is a checkerboard. If \code{checkerboards_only =
         TRUE}, then the random selection is constrained to existing
         checkerboards. In other words, if \code{checkerboards_only = FALSE}
         a lot of swaps will not result in an actual swap and hence the
         final matrix will be 'less' randomized.

         There is no accepted value to set \code{swaps}. One rule of thumb
         is to set it to double the number of 1's in the matrix (i.e.
         \code{sum(inmat) * 2}).

         The returned matrix will not have column and row names even if the
         input matrix had them.
}
\examples{
xdata <- matrix(ncol = 8, nrow = 5)
xdata[] <- sample(c(0, 1), length(xdata), TRUE)
res <- testingtravis:::shuffle_checkerboard(xdata, swaps = sum(xdata) * 2)
# row sums are constant
rowSums(xdata) == rowSums(res)
# column sums are constant too
colSums(xdata) == colSums(res)
}
\author{
Christof Neumann
}
