% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shuffle_rowwise}
\alias{shuffle_rowwise}
\title{Randomize entries in matrix while keeping row sums fixed}
\usage{
shuffle_rowwise(inmat)
}
\arguments{
\item{inmat}{numeric or integer matrix, 0/1 matrix}
}
\value{
a randomized matrix with the same dimensions as \code{inmat}, where
the sum of values within a row remains constant
}
\description{
This is a utility function and as such probably not of much interest for a
typical user.
}
\details{
The returned matrix will not have column and row names even if the
         input matrix had them.
}
\examples{
xdata <- matrix(ncol = 8, nrow = 5)
xdata[] <- sample(c(0, 1), length(xdata), TRUE)
res <- testingtravis:::shuffle_rowwise(xdata)
# row sums are constant
rowSums(xdata) == rowSums(res)
# column sums are not (although some may be, by chance)
colSums(xdata) == colSums(res)
}
\author{
Christof Neumann
}
