% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_facs.R
\name{simulate_facs}
\alias{simulate_facs}
\title{Simulate simple FACS data}
\usage{
simulate_facs(n_aus = 5, n_events = 200, n_contexts = 2, n_subjects = 10)
}
\arguments{
\item{n_aus}{numeric, number of action units (default is \code{5}). Can't be
larger than 26.}

\item{n_events}{numeric, number of events (default is \code{200})}

\item{n_contexts}{numeric, number of contexts (default is \code{2}). Must be
not larger than \code{n_aus}.}

\item{n_subjects}{numeric, number of subjects (default is \code{10})}
}
\value{
a \code{data.frame} with one row per event and one column for each
        AU, and columns for context, subject code and duration.
}
\description{
Simulate simple FACS data
}
\details{
The function generates data such that a given AU will appear most
         often in one context and rarely in other contexts.
         
         Subjects become relevant only for the durations of events, i.e. some
         subjects tend to have longer events while others have shorter 
         events (which is independent of context).
         
         It is possible that the results contain events without any AU 
         present.
}
\examples{
xdata <- simulate_facs(n_aus = 5,
                       n_events = 200,
                       n_contexts = 1,
                       n_subjects = 3)
head(xdata)
boxplot(xdata$duration ~ xdata$subject)
}
\author{
Christof Neumann
}
