% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sri_mat}
\alias{sri_mat}
\title{Simple ratio index}
\usage{
sri_mat(inmat)
}
\arguments{
\item{inmat}{integer matrix, rows represent observations/events and columns
individuals/units}
}
\value{
a square matrix with simple ratio index for pairs of individuals /
        units
}
\description{
This is a utility function and as such probably not of much interest for a
typical user.
}
\author{
Christof Neumann
}
